requirements.txt
loguru==0.7.0
environ==1.0
fastapi==0.101.1
uvicorn==0.23.2
requests==2.31.0
mysql-connector-python==8.1.0
SQLAlchemy==2.0.20
environs==9.5.0
PyJWT==2.8.0
pydantic==2.3.0
passlib==1.7.4